StravaWatch 1.0
This is a working Pebble watchface created using SDK 2.0. There are no configuration options at present and it relies on a PHP script not included in this repo. If you wish to use it then contact me and I will send you the PHP script used to scrape stats from a Strava public profile page.